Implemented import assets in content browser
git-svn-id: svn://192.168.0.101/eleot@233 7f037352-864a-4b62-bc5a-4f07d5e00ced

+++ Develompent/Source/Engine/Engine/Source/System/ConVar.cpp
@@ -166,11 +166,11 @@ void CConVar::SetValueBool( bool InValue )
 	SetMin( true, 0.f );
 	SetMax( true, 1.f );
 
-	if ( bHasMin && InValue < minVar )
+	if ( bHasMin && ( float )InValue < minVar )
 	{
 		*boolData = minVar;
 	}
-	else if ( bHasMax && InValue > maxVar )
+	else if ( bHasMax && ( float )InValue > maxVar )
 	{
 		*boolData = maxVar;
 	}
